Updated BlockDiagonalizationSolver to be able to handle Hamiltonians and eigenvector workspaces that hav more elements than contained in an unsigned int.

+++ Lib/src/Solvers/BlockDiagonalizationSolver.cpp
@@ -156,8 +156,8 @@ void BlockDiagonalizationSolver::init(){
 
 	//Calculate amount of memory required to store all the blocks of the
 	//Hamiltonian.
-	unsigned int hamiltonianSize = 0;
-	unsigned int eigenVectorsSize = 0;
+	unsigned long hamiltonianSize = 0;
+	unsigned long eigenVectorsSize = 0;
 	for(unsigned int n = 0; n < numStatesPerBlock.size(); n++){
 		hamiltonianSize += (numStatesPerBlock.at(n)*(numStatesPerBlock.at(n)+1))/2;
 		eigenVectorsSize += numStatesPerBlock.at(n)*numStatesPerBlock.at(n);
@@ -183,10 +183,10 @@ void BlockDiagonalizationSolver::init(){
 	}
 
 	if(getGlobalVerbose() && getVerbose()){
-		int numBytesHamiltonian = hamiltonianSize*sizeof(
+		long numBytesHamiltonian = hamiltonianSize*sizeof(
 			complex<double>
 		);
-		int numBytesEigenVectors = eigenVectorsSize*sizeof(
+		long numBytesEigenVectors = eigenVectorsSize*sizeof(
 			complex<double>
 		);
